requirements.txt
fastapi
uvicorn
stripe
python-dotenv
azure-functions
azure-core